[e2e] move CreateTestSessionOutputDir to common package
