+++ test/new-e2e/pkg/e2e/suite_utils.go
@@ -6,11 +6,9 @@
 package e2e
 
 import (
-	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
-	"time"
 
 	"testing"
 )
@@ -27,53 +25,6 @@ func (tl testLogger) Write(p []byte) (n int, err error) {
 	tl.t.Helper()
 	tl.t.Log(string(p))
 	return len(p), nil
-}
-
-// CreateTestSessionOutputDir creates and returns a directory for tests to store output files and artifacts.
-// A timestamp is included in the path to distinguish between multiple runs, and os.MkdirTemp() is
-// used to avoid name collisions between parallel runs.
-//
-// A new directory is created on each call to this function, it is recommended to save this result
-// and use it for all tests in a run. For example see BaseSuite.GetRootOutputDir().
-//
-// See runner.GetProfile().GetOutputDir() for the root output directory selection logic.
-//
-// See CreateTestOutputDir and BaseSuite.CreateTestOutputDir for a function that returns a subdirectory for a specific test.
-func CreateTestSessionOutputDir(outputRoot string) (string, error) {
-	// Append timestamp to distinguish between multiple runs
-	// Format: YYYY-MM-DD_HH-MM-SS
-	// Use a custom timestamp format because Windows paths can't contain ':' characters
-	// and we don't need the timezone information.
-	timePart := time.Now().Format("2006-01-02_15-04-05")
-	// create root directory
-	err := os.MkdirAll(outputRoot, 0755)
-	if err != nil {
-		return "", err
-	}
-	// Create final output directory
-	// Use MkdirTemp to avoid name collisions between parallel runs
-	outputRoot, err = os.MkdirTemp(outputRoot, fmt.Sprintf("%s_*", timePart))
-	if err != nil {
-		return "", err
-	}
-	if os.Getenv("CI") == "" {
-		// Create a symlink to the latest run for user convenience
-		// TODO: Is there a standard "ci" vs "local" check?
-		//       This code used to be in localProfile.GetOutputDir()
-		latestLink := filepath.Join(filepath.Dir(outputRoot), "latest")
-		// Remove the symlink if it already exists
-		if _, err := os.Lstat(latestLink); err == nil {
-			err = os.Remove(latestLink)
-			if err != nil {
-				return "", err
-			}
-		}
-		err = os.Symlink(outputRoot, latestLink)
-		if err != nil {
-			return "", err
-		}
-	}
-	return outputRoot, nil
 }
 
 // CreateTestOutputDir creates a directory for a specific test that can be used to store output files and artifacts.
